updated.
2002-05-28  Sven Neumann  <sven@gimp.org>

	* NEWS: updated.

	* plug-ins/common/jpeg.c (background_jpeg_save): cast stat.st_size
	to gint to avoid compiler warning with large file support. This
	will fail for JPEG files > 2GB but I doubt this is a real problem.

+++ plug-ins/common/jpeg.c
@@ -1129,8 +1129,8 @@ background_jpeg_save (PreviewPersistent *pp)
 
 	  stat (pp->file_name, &buf);
 	  g_snprintf (temp, sizeof (temp),
-		      _("Size: %lu bytes (%02.01f kB)"),
-		      buf.st_size, (float) (buf.st_size) / 1024.0);
+		      _("Size: %ld bytes (%02.01f kB)"),
+		      (gint) buf.st_size, (gdouble) (buf.st_size) / 1024.0);
 	  gtk_label_set_text (GTK_LABEL (preview_size), temp);
 	
 	  /* and load the preview */
